[HELIX-ENGINE]: Initial kernel integration
An effective solution to replace HTC's PNPMGR in Energy Aware Scheduling
kernels. Can be used on most devices even without EAS.

The kernel integration is only used to create new sysfs nodes under
/sys/power/helix_engine.

You need the binary included in my kernel in order to actually use the engine, and also have it set up the way it is in my kernel also.

Signed-off-by: ZeroInfinity <infinitezero345@gmail.com>